TrackRecordingService: re-worked TrackPoint recording tests for state changes.
